simplify docs and lock memcached version since we are locked to it with our hacks
