+++ README.md
@@ -1,42 +1,21 @@
 # LibmemcachedStore
 
-An ActiveSupport cache store that uses the C-based libmemcached client through Evan Weaver's Ruby/SWIG wrapper, [memcached](https://github.com/evan/memcached). libmemcached is fast (fastest memcache client for Ruby), lightweight, and supports consistent hashing, non-blocking IO, and graceful server failover.
-
-This cache is designed for Rails 3.2+ applications.
-
-## Prerequisites
-
-You'll need the memcached gem installed:
-
-```ruby
-gem install memcached
-```
-
-or in your Gemfile
-
-```ruby
-gem 'memcached'
-```
-
-There are no other dependencies.
+An ActiveSupport cache store that uses the fast, C-based, libmemcached client [memcached](https://github.com/evan/memcached).
+It is the fastest ruby memcached client, lightweight, supports consistent hashing, non-blocking IO, and graceful server failover.
 
 ## Installation
 
-Just add to your Gemfile
-
 ```ruby
-gem 'libmemcached_store', '~> 0.7.1'
+# Gemfile
+gem 'libmemcached_store'
 ```
-
-and you're set.
 
 ## Usage
 
-This is a drop-in replacement for the memcache store that ships with Rails. To
-enable, set the `config.cache_store` option to `libmemcached_store`
-in the config for your environment
+This is a drop-in replacement for the memcache store that ships with Rails.
 
 ```ruby
+# config/environments/*.rb
 config.cache_store = :libmemcached_store
 ```
 
@@ -51,20 +30,20 @@ config.cache_store = :libmemcached_store, %w(cache-01 cache-02 127.0.0.1:11212)
 Standard Rails cache store options can be used
 
 ```ruby
-config.cache_store = :libmemcached_store, '127.0.0.1:11211', {:compress => true, :expires_in => 3600}
+config.cache_store = :libmemcached_store, '127.0.0.1:11211', {compress: true, expires_in: 3600}
 ```
 
 More advanced options can be passed directly to the client
 
 ```ruby
-config.cache_store = :libmemcached_store, '127.0.0.1:11211', {:client => { :binary_protocol => true, :no_block => true }}
+config.cache_store = :libmemcached_store, '127.0.0.1:11211', {client: { binary_protocol: true, no_block: true }}
 ```
 
 You can also use `:libmemcached_store` to store your application sessions
 
 ```ruby
 require 'action_dispatch/session/libmemcached_store'
-config.session_store :libmemcached_store, :namespace => '_session', :expire_after => 1800
+config.session_store :libmemcached_store, namespace: '_session', expire_after: 1800
 ```
 
 You can use `:libmemcached_local_store` if you want a local in-memory cache for each request
@@ -77,7 +56,7 @@ config.cache_store :libmemcached_local_store
 Increment / Decrement only work on raw values:
 
 ```ruby
-Rails.cache.write 'x', '1', :raw => true
+Rails.cache.write 'x', '1', raw: true
 Rails.cache.increment 'x' # => 2
 Rails.cache.decrement 'x' # => 1
 ```
