+++ lib/libmemcached_store/version.rb
