+++ libmemcached_store.gemspec
@@ -1,4 +1,4 @@
-require "./lib/memcached/version"
+require "./lib/libmemcached_store/version"
 
 Gem::Specification.new do |s|
   s.name        = "libmemcached_store"
@@ -15,7 +15,7 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = '>= 2.2.0'
 
-  s.add_dependency('memcached')
+  s.add_dependency('memcached', '~> 1.8.0') # re-check lib/memcached/get_with_flags.rb before bumping
 
   s.add_development_dependency('rack')
   s.add_development_dependency('rake')
